Makes a scenery template given a block layer file
This file speeds up the process of making scenery by replacing all the
characters expect the '-' to hashTags. This makes the outline of the
blocks
easy to see and easier to place plants and fruits on.

NOTE: HashTags are read in by the Data file as empty sprites (Added by
Will).